tech changes
git-svn-id: svn://192.168.1.10/public/CoffeeMud@13717 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/BasicTech/GenSpaceShip.java
@@ -197,7 +197,7 @@ public class GenSpaceShip extends StdBoardable implements Electronics, SpaceShip
 							if(command==Technical.TechCommand.ACCELLLERATION)
 							{
 								final ThrustPort dir=(ThrustPort)parms[0];
-								final int amount=((Integer)parms[1]).intValue();
+								final double amount=((Double)parms[1]).doubleValue();
 								final boolean isConst = ((Boolean)parms[2]).booleanValue();
 								if(CMSecurity.isDebugging(DbgFlag.SPACESHIP))
 									Log.debugOut("SpaceShip "+name()+" accellerates "+amount+" to the "+dir.toString());
