+++ com/planet_ink/coffee_mud/Items/CompTech/StdShipFuellessThruster.java
@@ -46,7 +46,7 @@ public class StdShipFuellessThruster extends StdElecCompItem implements ShipComp
 
 	protected int		maxThrust		= 8900000;
 	protected int		minThrust		= 0;
-	protected int		thrust			= 0;
+	protected double	thrust			= 0;
 	protected long		specificImpulse	= SpaceObject.VELOCITY_SUBLIGHT;
 	protected double	fuelEfficiency	= 0.33;
 	protected boolean	constantThrust	= true;
@@ -99,13 +99,13 @@ public class StdShipFuellessThruster extends StdElecCompItem implements ShipComp
 	}
 
 	@Override
-	public int getThrust()
+	public double getThrust()
 	{
 		return thrust;
 	}
 
 	@Override
-	public void setThrust(int current)
+	public void setThrust(double current)
 	{
 		thrust = current;
 	}
